[Fix] 투표 결과 비율 계산에 double 타입 적용

+++ src/main/java/com/moa/moa_server/domain/vote/dto/response/VoteOptionResultWithId.java
@@ -4,5 +4,5 @@ public record VoteOptionResultWithId(
         Long voteId,
         int optionNumber,
         int count,
-        int ratio
+        double ratio
 ) {}
