+++ src/main/java/com/moa/moa_server/domain/vote/dto/response/result/VoteOptionResult.java
@@ -3,5 +3,5 @@ package com.moa.moa_server.domain.vote.dto.response.result;
 public record VoteOptionResult(
         int optionNumber,
         int count,
-        int ratio
+        double ratio
 ) {}
